src/test/java/RepoTests/ExpenseRepoTests/ExpenseCategoryRepoTests/CarRepoTests/PaymentsRepoImpTest.java
package RepoTests.ExpenseRepoTests.ExpenseCategoryRepoTests.CarRepoTests;

import BudgetBuddi.Domain.Expenses.ExpsenseCategories.Car.Payment;
import BudgetBuddi.Repository.ExpenseRepositories.ExpenseCategoryRepositories.CarRepositories.PayMentRepository;
import BudgetBuddi.Repository.Implentation.ExpensesRepoImplementation.ExpenseCategoryReposImplementaion.CarRepoImplementation.PaymentsRepoImp;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import java.util.ArrayList;

public class PaymentsRepoImpTest {

    private PayMentRepository repository;

    @Before
    public void setUp() throws Exception
    {
        this.repository = PaymentsRepoImp.getRepository();
    }

    @Test
    public void getRepository() {
    }

    @Test
    public void getAll()
    {
        ArrayList<Payment> payments = this.repository.getAll();
        Assert.assertEquals(1, payments.size());
    }

    @Test
    public void create()
    {
        this.repository.create(null);
        Assert.assertEquals(null, null);
    }

    @Test
    public void update() {
    }

    @Test
    public void delete()
    {
        this.repository.delete(null);
        Assert.assertEquals(null, null);
    }

    @Test
    public void read() {
    }
}